feat: consolidate music button handlers into a single module and refactor button exports
