+++ src/interactions/buttons/index.ts
@@ -4,11 +4,8 @@ import { handleRemoveRadio } from "./hope/handleRemoveRadio";
 import { handleMotusTry } from "./motus/handleMotusTry";
 import { handleQuizButton } from "./quiz/handleQuizButton";
 import { reportQuestionButton } from "./quiz/reportQuestionButton";
+import { backButton, loopButton, resumeAndPauseButton, skipButton } from "./music/musicButtons";
 
-import { backButton } from "../commands/music/back";
-import { loopButton } from "../commands/music/loop";
-import { resumeAndPauseButton } from "../commands/music/pause";
-import { skipButton } from "../commands/music/skip";
 
 export const buttons = {
     handleQuizButton,
