+++ src/interactions/commands/music/loop.ts
@@ -1,6 +1,6 @@
 import { errorEmbed, successEmbed } from "@/utils/embeds";
 import { QueueRepeatMode, useQueue } from "discord-player";
-import { ButtonInteraction, CommandInteraction, SlashCommandBuilder, SlashCommandOptionsOnlyBuilder } from "discord.js";
+import { CommandInteraction, SlashCommandBuilder, SlashCommandOptionsOnlyBuilder } from "discord.js";
 
 export const data: SlashCommandOptionsOnlyBuilder = new SlashCommandBuilder()
     .setName("loop")
@@ -61,19 +61,4 @@ export async function execute(interaction: CommandInteraction) {
             break
         }
     }
-}
-
-export async function loopButton(interaction: ButtonInteraction) {
-    const methods = ['disabled', 'track', 'queue']
-    const queue = useQueue(interaction.guildId as string)
-
-    if (!queue?.isPlaying()) return await interaction.reply({ embeds: [errorEmbed(interaction, new Error("Aucune musique n'est en cours de lecture."))], ephemeral: true })
-
-    if(queue.repeatMode === QueueRepeatMode.QUEUE) {
-        queue.setRepeatMode(QueueRepeatMode.OFF)
-    } else {
-        queue.setRepeatMode(queue.repeatMode + 1)
-    }
-
-    return await interaction.reply({ embeds: [successEmbed(interaction, `Boucle ${methods[queue.repeatMode]}`)], ephemeral: true })
 }
